feat : adding the possibility to link strava account during first connection

+++ lib/widgets/statistics/interactive_chart_widget.dart
@@ -9,7 +9,7 @@ class InteractiveChartWidget extends StatelessWidget {
 
   final int goalSteps;
   final int goalTime;
-  final double goalDistance;
+  final int goalDistance;
 
   final DateTime weekStartDate;
   final DateTime selectedDate;
